Доработал игровую логику и внёс небольшие улучшения по всему коду

+++ Assets/Scripts/Game Logic/Field.cs
@@ -6,116 +6,180 @@ public class Field : MonoBehaviour
     private List<List<Cell>> _cells;
     private GameParameters _parameters;
 
-    public IReadOnlyList<IReadOnlyList<Cell>> CellsData => _cells;
+    public IReadOnlyList<IReadOnlyList<Cell>> Cells => _cells;
     public GameParameters Parameters => _parameters;
 
-    public void SetField(List<List<Cell>> cellsData, GameParameters parameters)
+    public void SetField(List<List<Cell>> cells, GameParameters parameters)
     {
-        _cells = cellsData;
+        _cells = cells;
         _parameters = parameters;
     }
 
-    public bool TryMove(Vector2Int coordOut, Vector2Int coordIn)
+    public Vector2Int GetCoordinates(Cell cell)
+    {
+        if(cell == null)
+        {
+            throw new System.ArgumentNullException(cell.ToString());
+        }
+        for(int i = 0; i < _cells.Count; i++)
+        {
+            int index = _cells[i].FindIndex(item => item == cell);
+            if(index != -1)
+            {
+                return new Vector2Int(i, index);
+            }
+        }
+        throw new System.ArgumentException("The cell is not contained in the field");
+    }
+
+    public bool IsValideCoordinate(Vector2Int coordinate)
+    {
+        if(coordinate.x < 0 || coordinate.x >= _parameters.Width ||
+           coordinate.y < 0 || coordinate.y >= _parameters.Height)
+        {
+            return false;
+        }
+        if(_cells[coordinate.x][coordinate.y] == null)
+        {
+            return false;
+        }
+        return true;
+    }
+
+    public void OnSwipe(Cell cell, Vector2Int direction)
+    {
+        Vector2Int element1 = GetCoordinates(cell);
+        Vector2Int element2 = element1 + direction;
+        if(IsValideCoordinate(element2) == false)
+        {
+            return;
+        }
+        TryMove(element1, element2);
+    }
+
+    public void DeleteCell(Vector2Int coordinate)
+    {
+        if(IsValideCoordinate(coordinate) == false)
+        {
+            throw new System.ArgumentException("Invalid coordinates");
+        }
+        _cells[coordinate.x][coordinate.y].
+            Generate(_parameters.MinStartCellValue, _parameters.MaxStartCellValue);
+
+        if(_parameters.Mode == GameMode.Infinity || _parameters.Mode == GameMode.Level)
+        {
+            DeleteCells(GetDeleteMap());
+        }
+    }
+
+    private bool TryMove(Vector2Int coordOut, Vector2Int coordIn)
     {
         if(Mathf.Abs(coordIn.x - coordOut.x) + Mathf.Abs(coordIn.y - coordOut.y) != 1)
         {
             return false;
         }
-        if(_cells[coordIn.x][coordIn.y]?.Data.TryIncreaseValue(_cells[coordOut.x][coordOut.y]?.Data) != true)
+        if(_cells[coordIn.x][coordIn.y].Data.TryIncreaseValue(_cells[coordOut.x][coordOut.y].Data) != true)
         {
             return false;
         }
         DeleteCell(coordOut);
         if(_parameters.Mode == GameMode.Infinity || _parameters.Mode == GameMode.Level)
         {
-            DeleteCells(CheckFull());
+            DeleteCells(GetDeleteMap());
         }
         return true;
     }
 
-    private void DeleteCell(Vector2Int coord)
+    private void MoveCellUp(Vector2Int coordinate)
     {
-        _cells[coord.x][coord.y].Data.GenerateValue(_parameters.MinStartCellValue, _parameters.MaxStartCellValue);
-        _cells[coord.x][coord.y].State = Cell.ChangeState.Created;
-        for(int i = coord.y + 1; i < _cells[coord.x].Count; i++)
+        Cell rising = _cells[coordinate.x][coordinate.y];
+        Cell temp;
+        for(int i = coordinate.y + 1; i < _cells[coordinate.x].Count; i++)
         {
-            if(_cells[coord.x][i].State != Cell.ChangeState.Created)
+            if(_cells[coordinate.x][i] == null)
             {
-                _cells[coord.x][i].State = Cell.ChangeState.Fall;
+                continue;
             }
-            Swap(new Vector2Int(coord.x, i), new Vector2Int(coord.x, i - 1));
+            rising.Swap(_cells[coordinate.x][i]);
+            temp = rising;
+            rising = _cells[coordinate.x][i];
+            _cells[coordinate.x][i] = temp;
         }
-        
     }
 
-    private void Swap(Vector2Int coord1, Vector2Int coord2)
+    private bool[,] GetDeleteMap()
     {
-        GameObject temp = _cells[coord1.x][coord1.y].Parent; 
-        _cells[coord1.x][coord1.y].SetParent(_cells[coord2.x][coord2.y].Parent);
-        _cells[coord2.x][coord2.y].SetParent(temp);
-        Cell tempCell = _cells[coord1.x][coord1.y];
-        _cells[coord1.x][coord1.y] = _cells[coord2.x][coord2.y];
-        _cells[coord2.x][coord2.y] = tempCell;
-    }
-
-    private bool[,] CheckFull()
-    {
-        bool[,] deleteMap = new bool[_parameters.Width, _parameters.Height];
+        bool[,] horizontalDeleteMap = new bool[_parameters.Width, _parameters.Height];
+        bool[,] verticalDeleteMap = new bool[_parameters.Width, _parameters.Height];
         for(int i = 0; i < _parameters.Height; i++)
         {
-            CheckRow(deleteMap, i);
+            CheckRow(horizontalDeleteMap, i);
         }
         for(int i = 0; i < _parameters.Width; i++)
         {
-            CheckColumn(deleteMap, i);
+            CheckColumn(verticalDeleteMap, i);
         }
-        return deleteMap;
+        return UnionDeleteMap(horizontalDeleteMap, verticalDeleteMap);
     }
 
-    private void CheckRow(bool[,] deleteMap, int RowNumber)
+    private void CheckRow(bool[,] deleteMap, int rowNumber)
     {
-        int inRow = 0;
+        int counter = 0;
         for(int i = 0; i < _parameters.Width; i++)
         {
-            if(_cells[i][RowNumber].IsActive && _cells[i][RowNumber].Data.IsComplete)
+            if(_cells[i][rowNumber] != null && _cells[i][rowNumber].Data.IsComplete)
             {
-                inRow++;
+                counter++;
                 if(i < _parameters.Width - 1)
                 {
                     continue;
                 }
             }
-            if(inRow >= _parameters.FullInRow)
+            if(counter >= _parameters.FullInRow)
             {
-                for(int k = 0; k < inRow; k++)
+                for(int k = 0; k < counter; k++)
                 {
-                    deleteMap[i - k, RowNumber] = true;
+                    deleteMap[i - k, rowNumber] = true;
                 }
             }
+            counter = 0;
         }
     }
 
-    private void CheckColumn(bool[,] deleteMap, int ColumnNumber)
+    private void CheckColumn(bool[,] deleteMap, int columnNumber)
     {
-        int inColumn = 0;
+        int counter = 0;
         for(int i = 0; i < _parameters.Height; i++)
         {
-            if(_cells[ColumnNumber][i].IsActive && _cells[ColumnNumber][i].Data.IsComplete)
+            if(_cells[columnNumber][i] != null && _cells[columnNumber][i].Data.IsComplete)
             {
-                inColumn++;
+                counter++;
                 if(i < _parameters.Height - 1)
                 {
                     continue;
                 }
             }
-            if(inColumn >= _parameters.FullInRow)
+            if(counter >= _parameters.FullInColumn)
             {
-                for(int k = 0; k < inColumn; k++)
+                for(int k = 0; k < counter; k++)
                 {
-                    deleteMap[ColumnNumber, i - k] = true;
+                    deleteMap[columnNumber, i - k] = true;
                 }
             }
+            counter = 0;
         }
+    }
+
+    private bool[,] UnionDeleteMap(bool[,] map1, bool[,] map2)
+    {
+        for(int i = 0; i < map1.GetLength(0); i++)
+        {
+            for(int j = 0; j < map1.GetLength(1); j++)
+            {
+                map1[i, j] = map1[i, j] || map2[i, j];
+            }
+        }
+        return map1;
     }
 
     private void DeleteCells(bool[,] deleteMap)
@@ -124,7 +188,10 @@ public class Field : MonoBehaviour
         {
             for(int j = deleteMap.GetLength(1); j >= 0; j--)
             {
-                DeleteCell(new Vector2Int(i, j));
+                if(deleteMap[i, j] == true)
+                {
+                    DeleteCell(new Vector2Int(i, j));
+                }
             }
         }
     }
